backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
chardet==5.2.0
Django==5.0.6
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
et_xmlfile==2.0.0
inflection==0.5.1
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.2.2
rpds-py==0.27.1
scikit-learn==1.4.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
wheel==0.45.1
